initial comments from colton
