+++ src/sentry/api/api_owners.py
@@ -28,4 +28,4 @@ class ApiOwner(Enum):
     UNOWNED = "unowned"
     WEB_FRONTEND_SDKS = "team-web-sdk-frontend"
     GDX = "gdx"
-    EMERGE_TOOLS = "emerge-tool"
+    EMERGE_TOOLS = "emerge-tools"
